pyproject.toml
[project]
name = "cheddar-bench"
version = "0.1.0"
description = "Benchmark for evaluating CLI coding agents on bug detection"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15.0",
    "rich>=13.9.0",
    "pydantic>=2.10.0",
    "litellm>=1.0.0",
    "tenacity>=9.1.4",
    "json-repair>=0.30.0",
]

[project.optional-dependencies]
azure = ["azure-identity>=1.16.0"]

[project.scripts]
cheddar = "cheddar.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cheddar"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (needed for Typer)
]

[tool.ruff.lint.isort]
known-first-party = ["cheddar"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = ["azure.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnknownVariableType = "warning"  # Pydantic generic lists trigger false positives
include = ["src"]
exclude = ["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "azure-identity>=1.16.0",
    "mypy>=1.19.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
